livekit-test/requirements.txt
# VoxKit LiveKit Test — Python dependencies
# Install: pip install -r requirements.txt

# Core framework
livekit-agents>=1.4.0,<2.0.0

# Plugins
livekit-plugins-openai>=1.3.0
livekit-plugins-soniox>=1.3.0
livekit-plugins-silero>=1.3.0

# Environment
python-dotenv>=1.0.0
